+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 290421
+
++ Wrote tests for BehaviorTree class.
+
+## 120421
+
++ Wrote JavaDoc for classes from machine package.
+
+## 060421
+
++ Wrote tests for LeafNode, RootNode, SequenceNode, SelectorNode, InverterNode, LimitNode and RepeatNode classes.
++ Removed some "garbage" from machine package.
+
 ## 250321
 
 + Added class Initializer.